refactor: 폴더 위계 조정
