+++ src/components/Pages/Cart/Cart.stories.jsx
@@ -1,11 +1,15 @@
 import React from 'react';
-import Cart from 'components/Cart';
+import Cart from 'components/Pages/Cart';
 
 export default {
-  title: 'components/Cart',
+  title: 'components/Pages/Cart/Cart',
   component: Cart,
 };
 
 const Template = (args) => <Cart {...args} />;
 
 export const Default = Template.bind({});
+
+Default.args = {
+  cartList: [],
+};
