+++ src/components/common/Header/Header.stories.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Header from 'components/common/Header/index';
 
 export default {
-  title: 'components/Header',
+  title: 'components/common/Header',
   component: Header,
 };
