+++ src/components/common/Header/index.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { ReactComponent as CartIcon } from 'assets/icon/Cart.svg';
+
 import { ROUTE, COLOR } from 'constants';
+import { ReactComponent as CartIcon } from 'assets/icon/Cart.svg';
 
 const Styled = {
   Wrapper: styled.header`
